# WebApiComplexFood/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ProjectStructure.Api",
    "version": "v1"
  },
  "paths": {
    "/buyers/login": {
      "post": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerLoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BuyerDto"
            }
          }
        }
      }
    },
    "/buyers/{buyerId}": {
      "get": {
        "tags": [
          "Buyer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BuyerDto"
            }
          }
        }
      }
    },
...
    "/buyers/register": {
      "post": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerRegisterDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BuyerRegisterDto"
            }
          }
        }
      }
    },
    "/buyers/{id}": {
      "delete": {
        "tags": [
          "Buyer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/buyers/update/{buyerId}": {
      "put": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        }
      }
    },
    "/buyers/confirm/{buyerId}": {
      "patch": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerConfirmDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/buyers/deposit-balance/{buyerId}": {
      "patch": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerDepositBalanceDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/buyers/update-password/{buyerId}": {
      "patch": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerUpdatePasswordDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/buyers/forgot-password": {
      "post": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerForgotPasswordDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/buyers/change-password/{buyerId}": {
      "patch": {
        "tags": [
          "Buyer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyerChangePasswordDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/coupons/{buyerId}": {
      "get": {
        "tags": [
          "Coupon"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CouponDto"
              }
            }
          }
        }
      }
    },
    "/coupons/buy_coupons/{buyerId}": {
      "post": {
        "tags": [
          "Coupon"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CouponCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        }
      }
    },
...
    "/orders/{buyerId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderDto"
              }
            }
          }
        }
      }
    },
...
    "/orderItems/{orderId}": {
      "get": {
        "tags": [
          "OrderItem"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderItemDto"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Balance": {
      "type": "object",
      "properties": {
        "value": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "Buyer": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "$ref": "#/definitions/Email"
        },
        "password": {
          "$ref": "#/definitions/Password"
        },
        "firstName": {
          "$ref": "#/definitions/Name"
        },
        "lastName": {
          "$ref": "#/definitions/Name"
        },
        "phoneNumber": {
          "$ref": "#/definitions/PhoneNumber"
        },
        "gender": {
          "$ref": "#/definitions/Gender"
        },
        "confirmed": {
          "type": "boolean"
        },
        "balance": {
          "$ref": "#/definitions/Balance"
        },
        "confirmationCode": {
          "$ref": "#/definitions/UniqueCode"
        },
        "coupons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Coupon"
          }
        }
      }
    },
    "BuyerChangePasswordDto": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        }
      }
    },
    "BuyerConfirmDto": {
      "type": "object",
      "properties": {
        "confirmationCode": {
          "type": "string"
        }
      }
    },
    "BuyerDepositBalanceDto": {
      "type": "object",
      "properties": {
        "balance": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "BuyerDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "confirmed": {
          "type": "boolean"
        },
        "balance": {
          "format": "double",
          "type": "number"
        },
        "confirmationCode": {
          "type": "string"
        },
        "coupons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Coupon"
          }
        }
      }
    },
    "BuyerForgotPasswordDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "BuyerLoginDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "BuyerRegisterDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        }
      }
    },
    "BuyerUpdateDto": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      }
    },
    "BuyerUpdatePasswordDto": {
      "type": "object",
      "properties": {
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      }
    },
...
    "Categories": {
      "format": "int32",
      "enum": [
        1,
        2,
        3,
        4,
        5,
        6,
        7
      ],
      "type": "integer"
    },
    "Coupon": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "$ref": "#/definitions/TypeCoupons"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "code": {
          "$ref": "#/definitions/UniqueCode"
        },
        "buyer": {
          "$ref": "#/definitions/Buyer"
        },
        "buyerId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CouponCreateDto": {
      "type": "object",
      "properties": {
        "type": {
          "$ref": "#/definitions/TypeCoupons"
        }
      }
    },
    "CouponDto": {
      "type": "object",
      "properties": {
        "type": {
          "$ref": "#/definitions/TypeCoupons"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "buyerId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Email": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "Gender": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "Name": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "OrderDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        },
        "datePlaced": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        },
        "discount": {
          "format": "int32",
          "type": "integer"
        },
        "code": {
          "type": "string"
        },
        "buyerId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "OrderItemDto": {
      "type": "object",
      "properties": {
        "orderItemId": {
          "format": "int32",
          "type": "integer"
        },
        "cantity": {
          "format": "int32",
          "type": "integer"
        },
        "category": {
          "$ref": "#/definitions/Categories"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "OrderStatus": {
      "format": "int32",
      "enum": [
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "Password": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "PhoneNumber": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
...
    "TypeCoupons": {
      "format": "int32",
      "enum": [
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "UniqueCode": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    }
  }
}